Sof_ functions and workbooks used in AST assignment
These are functions and workbooks I used in my AST assignment. Some of it is more useful than others. The some_helper_functions help with controling and analysing the output from the batchrunner. The rest are in analysing the data out.
extra_functions can be called from any other file as long as you have it in the filepath and you import it.
main_analysis is the main analysis I did with a galaxy collision where gal1 had all 3 components but gal2 didnt have a disk (main_analysis_type_2 is for the same but where gal2 has all 3 components as well - there is a little bit changed with how the missing particles are dealt with).
visual_analysis is something which can be used to plot a heatmap based on the results - so it is possible to have an excel table, summarise the results and then adapt visual_analysis to provide a heatmap allowing visualisation.